Break out CLI tests per module file
